# internal/models/book.go
package models

import "gorm.io/gorm"

type Book struct {
	gorm.Model         // adds ID, created_at etc.
	Title       string `json:"title"`
	Author      string `json:"author"`
	Description string `json:"description"`
}

type BookResponse struct {
	Title       string `json:"title"`
	Author      string `json:"author"`
	Description string `json:"description"`
}